Super monster code crunch
* Abandon VS Code + paket + FAKE setup for Visual Studio 2017
* Re-work project structure to support other types of fronts
* Create slack bot that integrates with application ApplicationServices
* Implement a few new features

There is now enough here that you can actually start solving problems.
However, the event store is still in-memory, and it won't actually show who's
solved the problems, so not *quite* ready for production yet...

I hate really large commits like this one, but I've been in the zone for too
long to be able to split this sensibly. If anyone else but me was to review
the history, I'd have done it differently.

At least I'm going to tell myself so.